requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
safetensors>=0.4.0

# Training
pytorch-lightning>=2.1.0
wandb>=0.16.0
peft>=0.7.0  # LoRA desteği

# Image processing
opencv-python>=4.8.0
albumentations>=1.3.0
pillow>=10.0.0
rawpy>=0.18.0  # RAW image support

# Quantization & Export
onnx>=1.15.0
onnxruntime>=1.16.0
onnxruntime-gpu>=1.16.0
onnxsim>=0.4.33  # ONNX simplifier
ai-edge-torch>=0.1.0  # TFLite export
neural-compressor>=2.4.0  # Intel quantization

# Metrics & Evaluation
torchmetrics>=1.2.0
lpips>=0.1.4
pytorch-fid>=0.3.0

# Utilities
numpy>=1.24.0
tqdm>=4.66.0
pyyaml>=6.0.0
einops>=0.7.0
timm>=0.9.0

